Allow reuse of XmlDocumentFormatter instances
Create a separate format state object for each new call to
`XmlDocumentFormatter.format()` so that the `XmlDocumentFormatter`
object itself can be reused with the same configuration for formatting
multiple items, without bleeding state over.

This fixes a bug in which the depth parameter was not reset between
calls to `format()` when a malformed XML file that had an imbalanced
pair of open/close tags was skipped (configured to IGNORE or WARN).